Funcionalidade User/Admin implementada com validacao de permissao

+++ ProjetoIrineu/app/Http/Controllers/ExemptionController.php
@@ -6,7 +6,6 @@ namespace App\Http\Controllers;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Exemption;
 use Illuminate\Http\Request;
-
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use Redirect;
@@ -26,6 +25,9 @@ class ExemptionController extends Controller
     
     public function index()
     {
+        if (! Auth::user()->permission) {
+            return redirect()->route('home')->with('mensagem_aviso', 'Você não tem os privilegios para acessar esta pagina');
+        }
         $exemption = Exemption::all();
         return view('exemptions.list',compact('exemption'));
    
@@ -39,7 +41,9 @@ class ExemptionController extends Controller
      */
     public function create()
     { 
-        
+        if (! Auth::user()->permission) {
+            return redirect()->route('home')->with('mensagem_aviso', 'Você não tem os privilegios para acessar esta pagina');
+        }
     }
 
     /**
@@ -50,7 +54,9 @@ class ExemptionController extends Controller
      */
     public function store(Request $request)
     {
-      
+        if (! Auth::user()->permission) {
+            return redirect()->route('home')->with('mensagem_aviso', 'Você não tem os privilegios para acessar esta pagina');
+        }
             
     }
 
@@ -62,7 +68,9 @@ class ExemptionController extends Controller
      */
     public function show($id)
     {
-        //
+        if (! Auth::user()->permission) {
+            return redirect()->route('home')->with('mensagem_aviso', 'Você não tem os privilegios para acessar esta pagina');
+        }
     }
 
     /**
@@ -73,6 +81,9 @@ class ExemptionController extends Controller
      */
     public function edit($id)
     {
+        if (! Auth::user()->permission) {
+            return redirect()->route('home')->with('mensagem_aviso', 'Você não tem os privilegios para acessar esta pagina');
+        }
         $exemption = Exemption::findorfail($id);
         
         return view('exemptions.edit',['exemption'=> $exemption]);
@@ -88,6 +99,10 @@ class ExemptionController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (! Auth::user()->permission) {
+            return redirect()->route('home')->with('mensagem_aviso', 'Você não tem os privilegios para acessar esta pagina');
+        }
+
         $exemption = Exemption::findorfail($id);
 
         $exemption->update($request->all());
@@ -112,6 +127,11 @@ class ExemptionController extends Controller
      */
      public function destroy($id)
      {
+
+        if (! Auth::user()->permission) {
+            return redirect()->route('home')->with('mensagem_aviso', 'Você não tem os privilegios para acessar esta pagina');
+        }
+
         $exemption = Exemption::findorfail($id);
 
         $exemption->delete();
@@ -122,7 +142,10 @@ class ExemptionController extends Controller
 
     public function confirmDestroy($id)
     {
-        
+        if (! Auth::user()->permission) {
+            return redirect()->route('home')->with('mensagem_aviso', 'Você não tem os privilegios para acessar esta pagina');
+        }
+
         $exemption = Exemption::findorfail($id);
         
         return view('exemptions.deleteConfirm',['exemption'=> $exemption]);
